Feat: move main panel folder
