+++ src/app/my-page/interests/panel/InterestsContents.tsx
@@ -1,4 +1,4 @@
-import MainCard from '@/app/panel/MainCard'
+import MainCard from '@/app/panel/main-page/MainCard'
 import { ProjectType } from '@/app/panel/MainPage'
 import { IMainCard } from '@/types/IPostDetail'
 import { Box, CircularProgress, Grid } from '@mui/material'
