+++ src/app/panel/main-page/MainCarousel.tsx
